requirements.txt
Flask==3.0.3
gptpdf==0.0.12
Markdown==3.6
MarkupSafe==2.1.5
python-dotenv
openai